requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.8
PyPDF2==3.0.1
pdfplumber==0.10.3
Pillow==10.1.0
pdf2image==1.16.3
python-magic==0.4.27
aiofiles==23.2.1